separated out nao motion streamer, renamed some things
