+++ sic_framework/devices/common_naoqi/nao_motion_streamer.py
